Modelled JDAO
