+++ src/files.js
@@ -196,7 +196,9 @@ FOAM_FILES([
   { name: "foam/dao/GUIDDAO" },
   { name: "foam/dao/ReadOnlyDAO" },
   { name: "foam/dao/StoreAndForwardDAO" },
-  { name: "foam/dao/JDAO", flags: ['node'] },
+  { name: "foam/dao/Journal" },
+  { name: "foam/dao/JournalJava", flags: ['java'] },
+  { name: "foam/dao/JDAO", flags: ['web'] },
   { name: "foam/dao/JDAOJava", flags: ['java'] },
   { name: "foam/dao/Relationship" },
   { name: "foam/dao/RelationshipDAO" },
